+++ web/src/components/product-item/product-item.tsx
@@ -1,24 +1,18 @@
 import Link from 'next/link';
-import React, { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 
-import {
-  BodyCopyRegularSmall,
-  CaptionSmall,
-  LabelLargeBold,
-  LabelSmall,
-  StyledH3Title,
-} from '../../config/global-styled-components';
-import { colors, fonts } from '../../config/global-styles';
+import { CaptionSmall, LabelLargeBold, StyledH3Title } from '../../config/global-styled-components';
+import { colors } from '../../config/global-styles';
 import { Tags } from '../../model/filters/tags';
 
 import { Product } from '../../model/product';
 import Accordion from '../shared/accordion/accordion';
 
-import Chevron from '../shared/chevron';
 import PrimaryButton from '../shared/primary-button';
-import { calculateProductStock } from '../shared/utilities';
+
 import ProductItemCarousel from './product-item-carousel';
+import ProductItemSelectors from './product-item-selectors';
 
 const ProductItemContainer = styled.section`
   display: block;
@@ -62,13 +56,6 @@ const ProductItemDiscountPrice = styled(LabelLargeBold)`
   text-decoration-line: line-through;
 `;
 
-const ProductItemDescription = styled(BodyCopyRegularSmall)`
-  font-size: 16px;
-  line-height: 24px;
-  letter-spacing: 1px;
-  padding-top: 1rem;
-`;
-
 const TransparentBadge = styled(CaptionSmall)`
   display: flex;
   flex-direction: row;
@@ -78,52 +65,6 @@ const TransparentBadge = styled(CaptionSmall)`
 
   border-radius: 100px;
   border: 1px solid ${colors.ui.grey50percent};
-`;
-
-const ItemExtraQualities = styled.p`
-  font-family: ${fonts.primary};
-  font-style: normal;
-  font-weight: bold;
-  font-size: 12px;
-  line-height: 15px;
-  letter-spacing: 1px;
-  color: ${colors.ui.grey50percent};
-  text-transform: uppercase;
-`;
-
-const ItemExtraQualityRow = styled.div`
-  padding-top: 1rem;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-const FakeSelect = styled.div`
-  background: ${colors.ui.grey10percent};
-  border-radius: 8px;
-  width: 168px;
-  height: 42px;
-  border: ${colors.ui.grey10percent};
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0.8rem 1rem 0.8rem 1rem;
-`;
-
-const SizeTable = styled(LabelSmall)`
-  color: ${colors.ui.grey75percent};
-  text-decoration: underline;
-`;
-
-const StyledButton = styled.button`
-  border: none;
-`;
-
-const SizesTextContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
 `;
 
 const PaddedPrimaryButton = styled.div`
@@ -139,9 +80,6 @@ const DeliveryInformationContainer = styled.div`
 `;
 
 const ProductItemDisplay = ({ product }: { product: Product }) => {
-  const [quantity, setQuanity] = useState(0);
-  const [selectedSize, setSelectedSize] = useState('XS');
-
   return (
     <ProductItemContainer>
       <Link href="/categories/productos" passHref>
@@ -161,32 +99,7 @@ const ProductItemDisplay = ({ product }: { product: Product }) => {
         )}
         {product.tag !== Tags.Discount && <ProductItemPrice>${product.price}</ProductItemPrice>}
       </PriceDisplay>
-      <ProductItemDescription>{product.description}</ProductItemDescription>
-      <ItemExtraQualityRow>
-        <ItemExtraQualities>Cantidad</ItemExtraQualities>
-        <FakeSelect>
-          <StyledButton onClick={() => setQuanity(quantity - 1)}>-</StyledButton>
-          <LabelSmall>{quantity}</LabelSmall>
-          <StyledButton onClick={() => setQuanity(quantity + 1)}>+</StyledButton>
-        </FakeSelect>
-      </ItemExtraQualityRow>
-      {calculateProductStock(product) > 0 && (
-        <ItemExtraQualityRow>
-          <SizesTextContainer>
-            <ItemExtraQualities>Talle</ItemExtraQualities>
-            <Link href="adonde?" passHref>
-              <SizeTable>Ver tabla de talles</SizeTable>
-            </Link>
-          </SizesTextContainer>
-
-          <FakeSelect>
-            <LabelSmall>{selectedSize}</LabelSmall>
-            <StyledButton type="button" onClick={() => setSelectedSize('M')}>
-              <Chevron isOpen={false} />
-            </StyledButton>
-          </FakeSelect>
-        </ItemExtraQualityRow>
-      )}
+      <ProductItemSelectors product={product} />
       <ExtraPaddedPrimaryButton>
         <PrimaryButton text="Comprar Ahora" url="mercadolink" />
       </ExtraPaddedPrimaryButton>
